LayoutTests/platform/efl/fast/regions/get-region-flow-ranges-text-vert-lr-expected.txt

[CSSRegions] Test for Region::getRegionFlowRanges in vertical-rl mode. Text breaks into two regions

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Text node.
Flow test
PASS range1.length is 1
PASS Array [wrapper,0,#text,4] is equal to [wrapper,0,#text,4]
PASS range2.length is 1
PASS Array [#text,5,wrapper,1] is equal to [#text,5,wrapper,1]
PASS successfullyParsed is true

TEST COMPLETE
